requirements.txt


streamlit==1.44.1
pandas==2.2.0
pydeck==0.8.0
pyarrow>=15

pyrebase4==4.7.1
requests>=2.28,<3
